backend/requirements.txt
# Dependencias principales para el backend de Gestión de Carga de Trabajo
Django>=5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
